working soap interface to search for movies

+++ src/main/java/org/fhtech/yama/domain/NewMovie.java
@@ -7,8 +7,8 @@ import java.util.List;
 public class NewMovie {
     private String name;
     private String description;
-    private int duration;
-    private int releaseYear;
+    private short duration;
+    private short releaseYear;
     private List<Long> actors;
     private long studioId;
     private Genre genre;
